requirements.txt
Django>=3.0,<4.0
django_extra_fields==3.0.2
djangorestframework==3.13.1
python-dateutil==2.8.2
pandas==1.3.5